+++ SRC/tokenSystem.h
@@ -27,7 +27,7 @@ int BurstWord(char* word, int contractionStyle = 0);
 char* GetBurstWord(unsigned int n);
 char* JoinWords(unsigned int n,bool output = false);
 void ProcessSplitUnderscores();
-char* Tokenize(char* input,int& count,char** words,bool all = false, bool nomodify = false,bool oobstart = false);
+char* Tokenize(char* input,int& count,char** words,bool all = false,bool oobstart = false);
 int ValidPeriodToken(char* start, char* end, char next,char next2);
 char* ReadTokenMass(char* ptr, char* word);
 void ProcessSubstitutes();